Major update lab 2

+++ labb_2/shapes/Shape.h
@@ -4,14 +4,14 @@
 
 #ifndef LABB_2_SHAPE_H
 #define LABB_2_SHAPE_H
-
+#include "math.h"
 
 class Shape {
     char* color;
 public:
     static int numberOfObjects;
     Shape(char* color);
-    virtual int getArea() = 0;
+    virtual double getArea() = 0;
     virtual char* getColor();
     int getNumberOfObjects();
 };
